refactor: Add utilities for processing locust test results and generating comparison reports
