+++ .github/workflows/tests.yml
@@ -24,7 +24,7 @@ jobs:
       - name: Install Dagger CLI
         run: curl -fsSL https://dl.dagger.io/dagger/install.sh | BIN_DIR=/usr/local/bin sudo -E sh
       - name: Compare performance
-        run: dagger call test-performance >> pr_comment.md
+        run: dagger call test-performance --output-format=markdown >> pr_comment.md
       - name: Write PR comment
         uses: marocchino/sticky-pull-request-comment@v2
         with:
